requirements.txt
pytorch-pretrained-bert == 0.6.1
tqdm == 4.31.1
jupyter
tensorflow
tensorboardX
